requirements.txt
flask
cassandra-driver
redis
grift